Bean Validation

+++ src/main/java/com/fabiocanedo/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.fabiocanedo.dscommerce.services;
 import com.fabiocanedo.dscommerce.dto.ProductDTO;
 import com.fabiocanedo.dscommerce.entities.Product;
 import com.fabiocanedo.dscommerce.repositories.ProductRepository;
+import com.fabiocanedo.dscommerce.services.exceptions.DatabaseException;
+import com.fabiocanedo.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 
@@ -22,8 +27,7 @@ public class ProductService {
     public ProductDTO findById(Long id){
 
         Optional<Product> result = repository.findById(id);
-        Product product = result.get();
-
+        Product product = result.orElseThrow(()-> new ResourceNotFoundException("Recurso não encontrado"));
         return new ProductDTO(product);
     }
 
@@ -44,21 +48,30 @@ public class ProductService {
     }
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-
-        //Instancia o produto pelo ID
-        Product entity = repository.getReferenceById(id);
-        //Copia os dados do DTO para Entity
-        copyDtoToEntity(dto, entity);
-        //Salva no banco de dados
-        entity = repository.save(entity);
-
-        //Retorna o produto DTO
-        return new ProductDTO(entity);
+        try{
+            //Instancia o produto pelo ID
+            Product entity = repository.getReferenceById(id);
+            //Copia os dados do DTO para Entity
+            copyDtoToEntity(dto, entity);
+            //Salva no banco de dados
+            entity = repository.save(entity);
+            //Retorna o produto DTO
+            return new ProductDTO(entity);
+        }catch(EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void  delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try{
+            repository.deleteById(id);
+        }catch(DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
